KPI Crud

+++ src/main/java/com/spring/evalapi/service/KPIService.java
@@ -4,7 +4,8 @@ package com.spring.evalapi.service;
 import com.spring.evalapi.entity.Cycle;
 import com.spring.evalapi.entity.KPI;
 import com.spring.evalapi.repository.CycleRepository;
-import com.spring.evalapi.repository.KpiRepository;
+import com.spring.evalapi.repository.KPIRepository;
+import jakarta.validation.constraints.NotBlank;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,7 +16,7 @@ import java.util.Optional;
 public class KPIService {
 
     @Autowired
-    private KpiRepository kpiRepository;
+    private KPIRepository kpiRepository;
 
     @Autowired
     private CycleRepository  cycleRepository;
@@ -30,17 +31,33 @@ public class KPIService {
     }
 
 
-    public List<KPI> getKPIsByKPIProfileId(long kpiProfileId) {
-        return kpiRepository.findByKpiProfile_Id(kpiProfileId);
-    }
+//    public List<KPI> getKPIsByKPIProfileId(long kpiProfileId) {
+//        return kpiRepository.findByProfile_Id(kpiProfileId);
+//    }
 
     public KPI addKPI(KPI kpi) {
+        if (kpi.getCycle()!=null) {
+            Cycle cycle = cycleRepository.findById(kpi.getCycle().getId())
+                    .orElseThrow(() -> new IllegalArgumentException("Cycle not found !"));
+            kpi.setCycle(cycle);
+        }
         return kpiRepository.save(kpi);
     }
     public KPI updateKPI(KPI kpi) {
+        if (!kpiRepository.existsById(kpi.getId())){
+            throw new IllegalArgumentException("KPI  not Found");
+        }
+        if (kpi.getCycle()!=null) {
+            Cycle cycle = cycleRepository.findById(kpi.getCycle().getId())
+                    .orElseThrow(() -> new IllegalArgumentException("Cycle not found !"));
+            kpi.setCycle(cycle);
+        }
         return kpiRepository.save(kpi);
     }
     public void deleteKPI(long id) {
+        if (!kpiRepository.existsById(id)){
+            throw new IllegalArgumentException("KPI is Already not Found");
+        }
         kpiRepository.deleteById(id);
     }
